Wait until apply to create inspectpack daemon (#215)
This makes it so the dashboard plugin can be configured without creating the inspectpack daemon.  If the daemon is created in the constructor, the plugin cannot be conditionally applied.  Currently, if the plugin is configured and added as part of a development only config, the process running the production config will never exit.  With the changes here, the plugin can be configured but if it is never applied, the webpack process exits properly.

Fixes #96.
Fixes #125.